chore: create cli mod
